# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Nizi POS Connector API
  description: >-
    REST API for the Nizi POS Connector local service (UART display).
    All endpoints require the `X-API-Key` header and are intended for localhost use only.
  version: 1.0.0
servers:
  - url: http://127.0.0.1:9121
    description: Local Service

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key

security:
  - ApiKeyAuth: []

paths:
  /api/status:
    get:
      summary: Get device connection status
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  connected:
                    type: boolean
                  port:
                    type: string
                    nullable: true
                  device_id:
                    type: string
                    nullable: true
                  firmware_id:
                    type: string
                    nullable: true

  /api/connect:
    post:
      summary: Connect to device
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                port:
                  type: string
                  description: COM port or null for auto-detect
                  nullable: true
      responses:
        '200':
          description: OK

  /api/disconnect:
    post:
      summary: Disconnect device
      responses:
        '200':
          description: OK

  /api/command:
    post:
      summary: Send raw command
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [command]
              properties:
                command:
                  type: string
      responses:
        '200':
          description: OK

  /api/settings:
    post:
      summary: Update system settings
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                volume:
                  type: integer
                  minimum: 0
                  maximum: 100
                brightness:
                  type: integer
                  minimum: 0
                  maximum: 100
                screentime:
                  type: integer
                  minimum: 1
                  maximum: 3600
      responses:
        '200':
          description: OK

  /api/upload-image:
    post:
      summary: Upload JPEG image
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
                size:
                  type: string
                  default: "320x480"
      responses:
        '200':
          description: OK

  /api/firmware:
    get:
      summary: Get device firmware version and update URL
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  connected:
                    type: boolean
                  device_id:
                    type: string
                    nullable: true
                  firmware_id:
                    type: string
                    nullable: true
                  update_url:
                    type: string
                    nullable: true

  /api/timeout:
    post:
      summary: Set screen timeouts (QR and Pass/Fail)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                qr_timeout:
                  type: integer
                  minimum: 1
                  maximum: 3600
                  default: 300
                  description: QR screen timeout in seconds
                pf_timeout:
                  type: integer
                  minimum: 1
                  maximum: 3600
                  default: 20
                  description: Pass/Fail screen timeout in seconds
      responses:
        '200':
          description: OK

  /api/buzzer:
    post:
      summary: Enable or disable buzzer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                enabled:
                  type: integer
                  enum: [0, 1]
                  description: 1 to enable, 0 to disable. B30 has buzzer disabled by default.
      responses:
        '200':
          description: OK

  /api/idle-mode:
    post:
      summary: Set device idle/inactivity mode
      description: |
        Modes:
        - SINGLE: static background image (`image_name`).
        - CYCLE: alternate two images (`img1`, `time1`, `img2`, `time2`).
        - SLEEP: turn off backlight after inactivity (`sleep_ms`).
        - SLEEP_WAKE: auto-cycle Wake→Sleep (`wake_ms`, `sleep_ms`).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mode]
              properties:
                mode:
                  type: string
                  enum: [SINGLE, CYCLE, SLEEP, SLEEP_WAKE]
                image_name:
                  type: string
                  default: IMG1
                  description: Image to display — SINGLE mode only.
                img1:
                  type: string
                  default: IMG1
                  description: First image name — CYCLE mode.
                img2:
                  type: string
                  default: IMG2
                  description: Second image name — CYCLE mode.
                time1:
                  type: integer
                  default: 60000
                  description: Image 1 display duration in ms — CYCLE mode.
                time2:
                  type: integer
                  default: 60000
                  description: Image 2 display duration in ms — CYCLE mode.
                sleep_ms:
                  type: integer
                  default: 30000
                  description: Inactivity timeout before sleep (ms) — SLEEP and SLEEP_WAKE modes.
                wake_ms:
                  type: integer
                  default: 120000
                  description: Awake duration before next sleep (ms) — SLEEP_WAKE mode.
      responses:
        '200':
          description: OK

  /api/get-idle:
    post:
      summary: Query current idle configuration
      description: Sends GET_IDLE to the device and returns the result.
      responses:
        '200':
          description: OK

  /api/buzzer-test:
    post:
      summary: Trigger diagnostic buzzer test (BUZZERTEST)
      responses:
        '200':
          description: OK

  /api/ble:
    post:
      summary: Enable or disable Bluetooth (BLE_ON / BLE_OFF)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled:
                  type: boolean
                  description: true → BLE_ON, false → BLE_OFF
      responses:
        '200':
          description: OK

  /api/upload-idle-image:
    post:
      summary: Upload persistent idle image (IMG1 or IMG2)
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
                slot:
                  type: string
                  enum: ["1", "2"]
                  default: "1"
                  description: "Image slot: 1 for IMG1, 2 for IMG2"
                size:
                  type: string
                  default: "320x480"
      responses:
        '200':
          description: OK